Refactor.
